Änderung des Parameter auf Long

+++ inoUnitFakultät/ClsFunktion.vb
@@ -1,7 +1,8 @@
 ﻿Public Class ClsFunktion
-    Public Function Fakultaet(Eingabe As Short) As Short
-        Dim Ergebnis As Short
-        For I As Short = 1 To Eingabe
+    Public Function Fakultaet(Eingabe As Long) As Long
+        Dim Ergebnis As Long
+
+        For I As Long = 1 To Eingabe
             Ergebnis += I
         Next
         Return Ergebnis
